# src/main/resources/api.yaml
openapi: 3.0.0
info:
  title: Prova Azienda
  version: 1.0.0
paths:
 /employee:
   put:
      summary: Update an existing Employee
      description: Update an existing Employee by Id
      operationId: updateEmployee
      requestBody:
        description: Update an existent Employee 
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Employee'
          application/xml:
            schema:
              $ref: '#/components/schemas/Employee'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Employee'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Employee'
        '400':
          description: Invalid ID supplied
        '404':
          description: Employee not found
        '405':
          description: Validation exception
   post:
      
      summary: Add a new Employee 
      description: Add a new Employee 
      operationId: addEmployee
      requestBody:
        description: Create a new Employee
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Employee'
          application/xml:
            schema:
              $ref: '#/components/schemas/Employee'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Employee'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Employee'
        '405':
          description: Invalid input 
 /employee/{employeeId}:          
   get:
      summary: Find Employee by ID
      description: Returns a single Employee
      operationId: getEmployeeById
      parameters:
        - name: employeeId
          in: path
          description: ID of employee to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Employee'
        '400':
          description: Invalid ID supplied
        '404':
          description: Employee not found
   post:
      summary: Updates a employee  with form data
      description: ''
      operationId: updateEmployeeWithForm
      parameters:
        - name: employeeId
          in: path
          description: ID of employee that needs to be updated
          required: true
          schema:
            type: integer
            format: int64
        - name: name
          in: query
          description: Name of employee that needs to be updated
          schema:
            type: string
        - name: role
          in: query
          description: Role of employee that needs to be updated
          schema:
            type: string
      responses:
        '405':
          description: Invalid input        
   delete:
      summary: Deletes an employee
      description: delete an employee
      operationId: deleteEmployee
      parameters:
        - name: api_key
          in: header
          description: ''
          required: false
          schema:
            type: string
        - name: employeeId
          in: path
          description: Employee id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: Invalid employee value
...
components:
  schemas:
    Employee:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        role:
          type: string
...